PAPI PRESET events are defined in the API as int. Somehow, they were coming out of papiStdEventDefs.h as unsigned. By casting a #define explicitly to int, they now appear as signed values. This matches the API at the user level, but will inevitably throw warnings elsewhere.

+++ src/papiStdEventDefs.h
@@ -26,7 +26,7 @@ platform's documentation carefully.
    functions, etc.
 */
 
-#define PAPI_PRESET_MASK 0x80000000
+#define PAPI_PRESET_MASK ((int)0x80000000)
 #define PAPI_NATIVE_MASK 0x40000000
 #define PAPI_PRESET_AND_MASK 0x7FFFFFFF
 /*#define PAPI_NATIVE_AND_MASK 0x3FFFFFFF*/ /* this masks both native and preset bits */
